+++ packages/app-next-example-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # app-next-example-plugin
 
+## 0.0.17-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.16.0-next.2
+  - @backstage/frontend-plugin-api@0.9.1-next.2
+
 ## 0.0.17-next.1
 
 ### Patch Changes