+++ packages/backend-test-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/backend-test-utils
 
+## 1.1.0-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-events-node@0.4.5-next.3
+  - @backstage/backend-defaults@0.5.3-next.3
+  - @backstage/backend-app-api@1.0.2-next.2
+  - @backstage/backend-plugin-api@1.0.2-next.2
+  - @backstage/config@1.2.0
+  - @backstage/errors@1.2.4
+  - @backstage/types@1.1.1
+  - @backstage/plugin-auth-node@0.5.4-next.2
+
 ## 1.1.0-next.2
 
 ### Minor Changes