+++ plugins/events-node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-events-node
 
+## 0.4.5-next.3
+
+### Patch Changes
+
+- 9816f51: Add raw body information to `RequestDetails`
+  and use the raw body when validating incoming event requests.
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.0.2-next.2
+  - @backstage/errors@1.2.4
+  - @backstage/types@1.1.1
+
 ## 0.4.5-next.2
 
 ### Patch Changes